Corrected the extraction offset, in which a parameter like "l" can provide a delta offset such as "l+0.06", indicating that the extraction model has a length larger than the drawn device. Previously the value was assumed to be in microns but did not scale between the .ext file and the SPICE netlist.  Corrected so that it scales like the other parameter values, being converted to internal units and tracking the internal grid scale.

+++ extflat/EFbuild.c
@@ -736,15 +736,15 @@ efBuildDeviceParams(name, argc, argv)
 	    if ((offset = strchr(pptr + 1, '+')) != NULL)
 	    {
 		*offset = '\0';
-		newparm->parm_offset = atof(offset + 1);
+		newparm->parm_offset = atoi(offset + 1);
 	    }
 	    else if ((offset = strchr(pptr + 1, '-')) != NULL)
 	    {
 		*offset = '\0';
-		newparm->parm_offset = -atof(offset + 1);
+		newparm->parm_offset = -atoi(offset + 1);
 	    }
 	    else
-		newparm->parm_offset = 0.0;
+		newparm->parm_offset = 0;
 	}
 
 	// For parameters defined for cell defs, copy the whole
